requirements.txt
# Core Django and framework dependencies
Django>=5.2.8
djangorestframework>=3.16.1
django-cors-headers>=4.9.0

# Database dependencies
psycopg2-binary>=2.9.9
sqlparse>=0.5.3

# Caching and session dependencies  
django-redis>=6.0.0
redis>=7.0.1

# Task queue dependencies
celery>=5.5.3
kombu>=5.5.4
vine>=5.1.0
billiard>=4.2.2

# Environment configuration
python-decouple>=3.8

# HTTP requests for external APIs
requests>=2.31.0

# OpenAI integration (for AI services)
openai>=1.0.0

# Development dependencies
# Uncomment for development
# django-debug-toolbar>=4.2.0
# pytest>=7.4.0
# pytest-django>=4.7.0
# coverage>=7.3.0